% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.seedwords}
\alias{as.seedwords}
\title{Convinient function to convert a list to seed words}
\usage{
as.seedwords(x, upper = 1, lower = 2)
}
\arguments{
\item{x}{a list of characters vectors}

\item{upper}{index for seed words for higher scores}

\item{lower}{index for seed words for lower scores}
}
\description{
Convinient function to convert a list to seed words
}
