% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.textmodel2.R
\name{as.textmodel_lss.textmodel_wordvector}
\alias{as.textmodel_lss.textmodel_wordvector}
\title{Support wordvector version 0.5.1 or earlier}
\usage{
\method{as.textmodel_lss}{textmodel_wordvector}(
  x,
  seeds,
  terms = NULL,
  nested_weight = TRUE,
  verbose = FALSE,
  spatial = TRUE,
  ...
)
}
\description{
Support wordvector version 0.5.1 or earlier
}
\keyword{internal}
