% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_lss}
\alias{bootstrap_lss}
\title{[experimental] Compute polarity scores of words with different
hyper-parameters
A function to compute polarity scores of words by resampling
hyper-parameters from a fitted LSS model.}
\usage{
bootstrap_lss(
  x,
  what = c("seeds", "k"),
  output = c("terms", "coef"),
  by = 50,
  ...
)
}
\arguments{
\item{x}{a fitted textmodel_lss object.}

\item{what}{choose the parameter to resample in bootstrapping.}

\item{output}{choose the types of the output object. If \code{terms}, words are
sorted in descending order.}

\item{by}{passed to \code{seq()} to generate values for \code{k}; only used for when
\code{what = "k"}.}

\item{...}{additional arguments passed to \code{as.textmodel_lss()}.}
}
\description{
[experimental] Compute polarity scores of words with different
hyper-parameters
A function to compute polarity scores of words by resampling
hyper-parameters from a fitted LSS model.
}
