% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{calc_scores}
\alias{calc_scores}
\title{Internal function to calculate document score}
\usage{
calc_scores(mx, df_dic, score_only = FALSE)
}
\description{
Internal function to calculate document score
}

