% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{candidates}
\alias{candidates}
\title{[experimental] Find candidates for seed words through bootstrapping}
\usage{
candidates(x, rank = 10, ...)
}
\arguments{
\item{x}{a fitted textmodel_lss object.}

\item{rank}{the lowest rank of candidates returned from \code{\link[=bootstrap_lss]{bootstrap_lss()}}.}
}
\description{
[experimental] Find candidates for seed words through bootstrapping
}
\keyword{internal}
