% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{coef.textmodel_lss}
\alias{coef.textmodel_lss}
\alias{coefficients.textmodel_lss}
\title{Extract model coefficients from a fitted textmodel_lss object}
\usage{
\method{coef}{textmodel_lss}(object, ...)

coefficients.textmodel_lss(object, ...)
}
\arguments{
\item{object}{a fitted \link{textmodel_lss} object}

\item{...}{unused}
}
\description{
\code{coef()} extract model coefficients from a fitted \code{textmodel_lss}
object.  \code{coefficients()} is an alias.
}
\keyword{internal}
\keyword{textmodel}
