% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cohesion}
\alias{cohesion}
\title{Computes cohesion of components of latent semantic analysis}
\usage{
cohesion(x, bandwidth = 10)
}
\arguments{
\item{x}{a fitted \code{textmodel_lss}}

\item{bandwidth}{size of window for smoothing}
}
\description{
Computes cohesion of components of latent semantic analysis
}
