% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywords.R
\name{count_collocates}
\alias{count_collocates}
\title{Count collocation of target words}
\usage{
count_collocates(tokens, target, target_negative, window = 10,
  valuetype = "fixed", flag_target = FALSE)
}
\description{
Count collocation of target words
}
\examples{
count_collocates2(list(LETTERS, letters), 'C|J|o|w', window=2)
count_collocates2(list(LETTERS, letters), 'C|J|o|w', 'A|z', window=2)

microbenchmark::microbenchmark(
count_collocates2(list(LETTERS, letters), 'C|J|o|w', window=2),
count_collocates(list(LETTERS, letters), 'C|J|o|w', window=2))
}

