% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_ideology}
\alias{data_dictionary_ideology}
\title{Seed words for analysis of left-right political ideology}
\description{
Seed words for analysis of left-right political ideology
}
\examples{
as.seedwords(data_dictionary_ideology)
}
