% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_sentiment}
\alias{data_dictionary_sentiment}
\title{Seed words for analysis of positive-negative sentiment}
\description{
Seed words for analysis of positive-negative sentiment
}
\examples{
as.seedwords(data_dictionary_sentiment)
}
\references{
Turney, P. D., & Littman, M. L. (2003). Measuring Praise and
  Criticism: Inference of Semantic Orientation from Association. ACM Trans.
  Inf. Syst., 21(4), 315–346. https://doi.org/10.1145/944012.944013
}
