% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_textmodel_lss_russianprotests}
\alias{data_textmodel_lss_russianprotests}
\title{A fitted LSS model on street protest in Russia}
\description{
This model was trained on a Russian media corpus (newspapers, TV transcripts
and newswires) to analyze framing of street protests. The scale is protests
as "freedom of expression" (high) vs "social disorder" (low). Although some
slots are missing in this object (because the model was imported from the
original Python implementation), it allows you to scale texts using
\code{predict}.
}
\references{
Lankina, Tomila, and Kohei Watanabe. “'Russian Spring' or 'Spring
Betrayal'? The Media as a Mirror of Putin's Evolving Strategy in Ukraine.”
Europe-Asia Studies 69, no. 10 (2017): 1526–56.
https://doi.org/10.1080/09668136.2017.1397603.
}
\keyword{data}
