% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagnosys}
\alias{diagnosys}
\title{Identify noisy documents in a corpus}
\usage{
diagnosys(x, ...)
}
\arguments{
\item{x}{character or \link{corpus} object whose texts will be diagnosed}

\item{...}{extra arguments passed to \code{tokens}}
}
\description{
Identify noisy documents in a corpus
}
\keyword{internal}
