% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrimination}
\alias{discrimination}
\title{Function to test qualtiy of seed words}
\usage{
discrimination(object, newdata = NULL)
}
\description{
Computes kurtosis of document scores and terms weights
}
\keyword{internal}
