% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divergence}
\alias{divergence}
\title{Function to test qualtiy of word vectors
Computes similarity between and within seed words}
\usage{
divergence(object)
}
\arguments{
\item{object}{a fitted LSS textmodel}
}
\description{
Function to test qualtiy of word vectors
Computes similarity between and within seed words
}
\keyword{internal}
