% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywords.R
\name{findCollocates}
\alias{findCollocates}
\title{Find significant collocations of targets}
\usage{
findCollocates(tokens, target, target_negative, window = 10, count_min = 5,
  word_only = TRUE, p = 0.001, ...)
}
\description{
Find significant collocations of targets
}
\examples{
docs <- readLines('/home/kohei/projects/immigration/data/uk_img/2009-2010.txt')
sents <- tokenize(docs, what='sentence', simplify = TRUE)
tokens <- tokenize(sents, removePunct=TRUE, removeNumbers=TRUE)
entries <- selectEntrywords(tokens, '^(immigra|migra)')


}

