% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedwords.R
\name{findSeeds}
\alias{findSeeds}
\title{Automatically select seed words}
\usage{
findSeeds(mx, words, mx_doc, scores, candidates, top, method = "pearson",
  min = 10, max = 100, tol = 0.001, dist = 0, power = 1,
  zigzag = FALSE, plot = FALSE)
}
\arguments{
\item{mx}{SVD-decomposed dfm}

\item{words}{entry words to dictionary}

\item{mx_doc}{dfm of training documents}

\item{score}{scores for training documents}

\item{candiates}{candiate words for seeds (entry words are used if missing)}
}
\description{
Automatically select seed words
}

