% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{get_beta}
\alias{get_beta}
\title{Internal function to beta parameters}
\usage{
get_beta(x, y, feature = NULL, method = "cosine")
}
\arguments{
\item{x}{svd-reduced dfm}

\item{y}{named-numberic vector for seed words}

\item{feature}{feature for which beta will be calcualted}
}
\description{
Internal function to beta parameters
}
\keyword{internal}
