% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_fixed_seeds}
\alias{get_fixed_seeds}
\title{Internal function to convert seed words from glob to fiexd format}
\usage{
get_fixed_seeds(seeds, types)
}
\description{
Internal function to convert seed words from glob to fiexd format
}

