% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{get_params}
\alias{get_params}
\title{Internal function to compute parameters}
\usage{
get_params(x, y, feature = NULL, method = "cosine")
}
\arguments{
\item{x}{svd-reduced dfm}

\item{y}{named-numberic vector for seed words}

\item{feature}{feature for which beta will be calcualted}
}
\description{
Internal function to compute parameters
}
\keyword{internal}
