% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywords.R
\name{gscore}
\alias{gscore}
\title{Internal function to calcualte g-score (Chi-square at the moment)}
\usage{
gscore(col, non, sum_col, sum_non, smooth = 1)
}
\description{
Internal function to calcualte g-score (Chi-square at the moment)
}

