% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{makeDictionary}
\alias{makeDictionary}
\title{Make a LSS dicitonary with seed words}
\usage{
makeDictionary(mx, words, seeds, valuetype = "fixed")
}
\description{
Make a LSS dicitonary with seed words
}

