% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{predict.textmodel_lss_fitted}
\alias{predict.textmodel_lss_fitted}
\title{prediction method for textmodel_lss}
\usage{
\method{predict}{textmodel_lss_fitted}(object, newdata = NULL,
  confidence.fit = FALSE)
}
\arguments{
\item{object}{a fitted LSS textmodel}

\item{newdata}{dfm on which prediction should be made}

\item{confidence.fit}{if \code{TRUE}, it also returns standard error of document scores.}
}
\description{
prediction method for textmodel_lss
}
