% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{textplot_scale1d}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{quanteda}{\code{\link[quanteda]{textplot_scale1d}}}
}}

