% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowMaxs}
\alias{rowMaxs}
\title{Maximum values in each row of a matrix}
\usage{
rowMaxs(x, absolute = FALSE)
}
\arguments{
\item{x}{a matrix object.}

\item{absolute}{if \code{TRUE}, return largest positive or negative values.}
}
\description{
Maximum values in each row of a matrix
}
\keyword{internal}
