% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{similarity}
\alias{similarity}
\title{Calcualte cosine similarities of words}
\usage{
similarity(mx, words, seeds)
}
\description{
Calcualte cosine similarities of words
}
\examples{
docs <- readLines('/home/kohei/projects/immigration/data/uk_img/2009-2010.txt')
mx <- quanteda::dfm(docs)
mx_lsa <- decompose(mx)
mx_sim <- similarity(mx_lsa)
}

