% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strength}
\alias{strength}
\title{Experimental function to test quality of seed words}
\usage{
strength(object)
}
\arguments{
\item{object}{a fitted \code{textmodel_lss}}
}
\description{
Experimental function to test quality of seed words
}
\keyword{internal}
