% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_plss.R
\name{textmodel_plss}
\alias{textmodel_plss}
\title{[experimental] Fit a Probabilistic Latent Semantic Scaling model}
\usage{
textmodel_plss(
  x,
  seeds,
  terms = NULL,
  k = 200,
  engine = c("word2vec"),
  tolower = TRUE,
  include_data = FALSE,
  group_data = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a \link[quanteda:tokens]{quanteda::tokens} or \link[quanteda:tokens_xptr]{quanteda::tokens_xptr} object.}

\item{seeds}{a character vector or named numeric vector that contains seed
words. If seed words contain "*", they are interpreted as glob patterns.
See \link[quanteda:valuetype]{quanteda::valuetype}.}

\item{terms}{a character vector or named numeric vector that specify words
for which polarity scores will be computed; if a numeric vector, words' polarity
scores will be weighted accordingly; if \code{NULL}, all the features of
\code{\link[quanteda:dfm]{quanteda::dfm()}} or \code{\link[quanteda:fcm]{quanteda::fcm()}} will be used.}

\item{k}{the number of singular values requested to the SVD engine. Only used
when \code{x} is a \code{dfm}.}

\item{engine}{currently only supports the word2vec model with the skip-gram algorithm.}

\item{tolower}{if \code{TRUE}, lower-case all the words in the model.}

\item{include_data}{if \code{TRUE}, fitted model includes the dfm supplied as \code{x}.}

\item{group_data}{if \code{TRUE}, apply \code{dfm_group(x)} before saving the dfm.}

\item{verbose}{show messages if \code{TRUE}.}

\item{...}{additional arguments passed to the underlying engine.}
}
\description{
Probabilistic Latent Semantic Scaling (PLSS) is a semi-supervised algorithm
document scaling based on language models.
}
