% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot_components}
\alias{textplot_components}
\title{[experimental] Plot clusters of word vectors}
\usage{
textplot_components(
  x,
  n = 5,
  method = "ward.D2",
  scale = c("absolute", "relative")
)
}
\arguments{
\item{x}{a fitted \code{textmodel_lss}}

\item{n}{the number of cluster}

\item{method}{the method for hierarchical clustering}

\item{scale}{change the scale of y-axis}
}
\description{
Experimental function to find clusters of word vectors
}
