% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot_factor}
\alias{textplot_factor}
\title{Plot factors of latent semanitic space}
\usage{
textplot_factor(x)
}
\arguments{
\item{x}{fitted textmodel_lss object}
}
\description{
Plot factors of latent semanitic space
}
