% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{textplot_proxy}
\alias{textplot_proxy}
\title{Plot proximity of seed words}
\usage{
textplot_proxy(x, group = TRUE)
}
\arguments{
\item{x}{fitted textmodel_lss object}

\item{group}{if \code{TRUE} group seed words by seed patterns and show
average proximity}
}
\description{
Plot proximity of seed words
}
