% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot_scale1d_features}
\alias{textplot_scale1d_features}
\title{compied from quanteda::textplot_scale1d_features}
\usage{
textplot_scale1d_features(
  x,
  weight,
  featlabels,
  highlighted = NULL,
  alpha = 0.7,
  highlighted_color = "black"
)
}
\description{
compied from quanteda::textplot_scale1d_features
}
