% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot_terms}
\alias{textplot_terms}
\title{Plot polarity scores of words}
\usage{
textplot_terms(x, highlighted = NULL)
}
\arguments{
\item{x}{fitted textmodel_lss object}

\item{highlighted}{a character vector to specify words to be highlighted}
}
\description{
Plot polarity scores of words
}
