% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.seedwords}
\alias{as.seedwords}
\title{Convinient function to convert a list to seed words}
\usage{
as.seedwords(x, upper = 1, lower = 2)
}
\arguments{
\item{x}{a list of characters vectors or a \link[quanteda:dictionary]{dictionary} object}

\item{upper}{numeric index or key for seed words for higher scores}

\item{lower}{numeric index or key for seed words for lower scores}
}
\value{
named numeric vector for seed words with polarity scores
}
\description{
Convinient function to convert a list to seed words
}
