% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{as.textmodel_lss}
\alias{as.textmodel_lss}
\title{Create a dummy textmodel_lss object from numeric vector}
\usage{
as.textmodel_lss(x)
}
\arguments{
\item{x}{named numeric vector}
}
\value{
a textmodel_lss object with \code{x} as polarity words
}
\description{
Create a dummy textmodel_lss object from numeric vector
}
\examples{
v <- c("a" = 0.1, "z" = -0.2, "d" = 0.3, "h" = -0.05)
lss <- as.textmodel_lss(v)
}
\keyword{internal}
