% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{as.textmodel_lss}
\alias{as.textmodel_lss}
\title{Create a dummy textmodel_lss object from numeric vector}
\usage{
as.textmodel_lss(x)
}
\arguments{
\item{x}{named numeric vector}
}
\description{
Create a dummy textmodel_lss object from numeric vector
}
\keyword{internal}
