% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_lss}
\alias{bootstrap_lss}
\title{[experimental] Compute polarity scores with different hyper-parameters}
\usage{
bootstrap_lss(
  x,
  what = c("seeds", "k"),
  mode = c("terms", "coef"),
  remove = FALSE,
  from = 50,
  to = NULL,
  by = 50,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a fitted textmodel_lss object.}

\item{what}{choose the hyper-parameter to resample in bootstrapping.}

\item{mode}{choose the type of the result of bootstrapping. If \code{coef},
returns polarity scores; if \code{terms}, returns words sorted by
the polarity scores in descending order.}

\item{remove}{if \code{TRUE}, remove each seed word when \code{what = "seeds"}.}

\item{from, to, by}{passed to \code{seq()} to generate values for \code{k}; only used
when \code{what = "k"}.}

\item{verbose}{show messages if \code{TRUE}.}

\item{...}{additional arguments passed to \code{as.textmodel_lss()}.}
}
\description{
A function to compute polarity scores of words by resampling hyper-parameters
from a fitted LSS model.
}
