% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_lss}
\alias{bootstrap_lss}
\title{[experimental] Compute polarity scores of words with different hyper-parameters.}
\usage{
bootstrap_lss(
  x,
  what = c("seeds", "k", "slice"),
  by = 50,
  n = 10,
  size = 100,
  ...
)
}
\arguments{
\item{x}{a fitted textmodel_lss object.}

\item{what}{choose the parameter to resample in bootstrapping.}

\item{by}{only for "k"}

\item{n}{only for "slice"}

\item{size}{only for "slice"}
}
\description{
[experimental] Compute polarity scores of words with different hyper-parameters.
}
\keyword{internal}
