% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boundary}
\alias{boundary}
\title{[experimental] Find clusters of word vectors}
\usage{
boundary(x, n = 10, method = "ward.D2")
}
\arguments{
\item{x}{a fitted \code{textmodel_lss}}

\item{n}{the number of cluster}

\item{method}{the method for hierarchical clustering}
}
\description{
Experimental function to find clusters of word vectors
}
\keyword{internal}
