% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{char_keyness}
\alias{char_keyness}
\title{identify words strongly associated with target word based on collocation}
\usage{
char_keyness(x, pattern, window = 10, p = 0.001, min_count = 10, ...)
}
\arguments{
\item{x}{tokens object created by \code{\link[quanteda]{tokens}}.}

\item{window}{size of window for collocation analysis.}

\item{p}{threashold for statistical significance of collocaitons.}

\item{min_count}{minimum frequency for words within the window to be
considered as collocations.}

\item{...}{additional arguments passed to \code{\link{textstat_keyness}}.}
}
\description{
identify words strongly associated with target word based on collocation
}
\examples{
require(quanteda)
load('/home/kohei/Dropbox/Public/guardian-sample.RData')
corp <- corpus_reshape(data_corpus_guardian, 'sentences')
toks <- tokens(corp, remove_punct = TRUE)
toks <- tokens_remove(toks, stopwords())

# economy keywords
eco <- char_keyness(toks, 'econom*')
head(eco)

# politics keywords
pol <- char_keyness(toks, 'politi*')
head(pol)
}
\seealso{
\code{\link{textstat_keyness}}
}
