% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cohesion}
\alias{cohesion}
\title{\link{Experimental} Computes cohesion of components of latent semantic analysis}
\usage{
cohesion(object, bandwidth = 10)
}
\arguments{
\item{object}{a fitted \code{textmodel_lss}}

\item{bandwidth}{size of window for smoothing}
}
\description{
\link{Experimental} Computes cohesion of components of latent semantic analysis
}
