% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrimination}
\alias{discrimination}
\alias{strength}
\title{Function to test quality of seed words}
\usage{
discrimination(object, newdata = NULL)

strength(object)
}
\description{
Gauge seed words' ability to discriminate unit texts by computing kurtosis of
document scores and terms weights.
}
\keyword{internal}
