% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{optimize_lss}
\alias{optimize_lss}
\title{[experimental] Compute variance ratio with different hyper-parameters}
\usage{
optimize_lss(x, ...)
}
\arguments{
\item{x}{a fitted textmodel_lss object.}

\item{...}{additional arguments passed to \code{\link[=as.textmodel_lss]{as.textmodel_lss()}} and
\code{\link[=predict]{predict()}}.}
}
\description{
[experimental] Compute variance ratio with different hyper-parameters
}
\keyword{internal}
