% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{predict.textmodel_lss}
\alias{predict.textmodel_lss}
\title{prediction method for textmodel_lss}
\usage{
\method{predict}{textmodel_lss}(object, newdata = NULL, fit.se = FALSE,
  density = FALSE, rescaling = TRUE, ...)
}
\arguments{
\item{object}{a fitted LSS textmodel}

\item{newdata}{dfm on which prediction should be made}

\item{fit.se}{if \code{TRUE}, it returns standard error of document scores.}

\item{density}{if \code{TRUE}, returns frequency of features in documents.
Density distribution of features can be used to remove documents about
unrelated subjects.}

\item{rescaling}{if \code{TRUE}, scores are reslaced using \code{scale()}.}

\item{...}{not used}
}
\description{
prediction method for textmodel_lss
}
