% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seedwords}
\alias{seedwords}
\title{Seed words for Latent Semantci Analysis}
\usage{
seedwords(type)
}
\arguments{
\item{type}{type of seed words currently only for sentiment (\code{sentiment})
or political ideology (\code{ideology}).}
}
\description{
Seed words for Latent Semantci Analysis
}
\examples{
seedwords('sentiment')
}
\references{
Turney, P. D., & Littman, M. L. (2003). Measuring Praise and
Criticism: Inference of Semantic Orientation from Association. ACM Trans.
Inf. Syst., 21(4), 315–346. https://doi.org/10.1145/944012.944013
}
