% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{seedwords}
\alias{seedwords}
\title{Seed words for sentiment analysis}
\usage{
seedwords(type)
}
\arguments{
\item{type}{type of seed words currently only for sentiment (\code{pos-neg})
or political ideology (\code{left-right}).}
}
\description{
Seed words for sentiment analysis
}
\examples{
seedwords('pos-neg')
}
\references{
Turney, P. D., & Littman, M. L. (2003). Measuring Praise and
  Criticism: Inference of Semantic Orientation from Association. ACM Trans.
  Inf. Syst., 21(4), 315–346. https://doi.org/10.1145/944012.944013
}
