% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{smooth_lss}
\alias{smooth_lss}
\title{Smooth predicted LSS scores by local polynomial regression}
\usage{
smooth_lss(x, lss_var = "fit", date_var = "date", span = 0.1,
  from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{a \code{data.frame} containing variables for LSS scores and dates}

\item{lss_var}{the name of the column for LSS scores}

\item{date_var}{the name of the columns for dates}

\item{span}{determines the level of smoothing}

\item{from}{start of the time period}

\item{to}{end of the time period}

\item{...}{extra arguments passed to \code{\link{loess}}}
}
\description{
Smooth predicted LSS scores by local polynomial regression
}
