% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{textmodel_lss}
\alias{textmodel_lss}
\title{a vector-space model for subject specific sentiment-analysis}
\usage{
textmodel_lss(x, y, pattern = NULL, k = 300, verbose = FALSE,
  cache = FALSE, ...)
}
\arguments{
\item{x}{a dfm created by \code{\link[quanteda]{dfm}}}

\item{y}{character vector or named character vector that contains seed words.}

\item{pattern}{pattern to select featues to make models only sensitive to
subject specific words.}

\item{k}{the size of semantic space passed to \code{\link[RSpectra]{svds}}}

\item{cache}{if \code{TRUE}, save retult of SVD for next execution with
identical \code{x} and \code{k}.}

\item{...}{additional argument passed to \code{\link[RSpectra]{svds}}}
}
\description{
a vector-space model for subject specific sentiment-analysis
}
\examples{
require(quanteda)

load('/home/kohei/Dropbox/Public/guardian-sample.RData')
corp <- corpus_reshape(data_corpus_guardian, 'sentences')
toks <- tokens(corp)
mt <- dfm(toks, remove = stopwords())
mt <- dfm_trim(mt, min_count = 10)
lss <- textmodel_lss(mt, seedwords('pos-neg'))

# sentiment model on economy
eco <- char_keyness(toks, 'econom*')
lss_eco <- textmodel_lss(mt, seedwords('pos-neg'), pattern = eco)

# sentiment model on politics
pol <- char_keyness(toks, 'polti*')
lss_pol <- textmodel_lss(mt, seedwords('pos-neg'), pattern = pol)
}
\references{
Watanabe, Kohei. “Measuring News Bias: Russia’s Official News
  Agency ITAR-TASS’ Coverage of the Ukraine Crisis.” European Journal of
  Communication 32, no. 3 (March 20, 2017): 224–41.
  https://doi.org/10.1177/0267323117695735.
}
