% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{textmodel_lss}
\alias{textmodel_lss}
\title{A vector-space model for subject specific sentiment-analysis}
\usage{
textmodel_lss(x, seeds, features = NULL, k = 300, cache = FALSE,
  simil_method = "cosine", include_data = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{a dfm created by \code{\link[quanteda]{dfm}}}

\item{seeds}{a character vector, named numeric vector or dictionary that contains
seed words.}

\item{features}{featues of a dfm to be included in the model as terms. This
argument is used to make models only sensitive to subject specific words.}

\item{k}{the size of semantic space passed to \code{\link[RSpectra]{svds}}}

\item{cache}{if \code{TRUE}, save retult of SVD for next execution with
identical \code{x} and \code{k}.}

\item{simil_method}{specifies method to compute similiaty between features.
The value is passed to \code{\link[quanteda]{textstat_simil}}, "cosine" is
used otherwise.}

\item{include_data}{if \code{TRUE}, fitted model include the dfm supplied as
\code{x}.}

\item{verbose}{show messages if \code{TRUE}.}

\item{...}{additional argument passed to \code{\link[RSpectra]{svds}}}
}
\description{
A vector-space model for subject specific sentiment-analysis
}
\examples{
require(quanteda)

load('/home/kohei/Dropbox/Public/guardian-sample.RData')
corp <- corpus_reshape(data_corpus_guardian, 'sentences')
toks <- tokens(corp, remove_punct = TRUE)
mt <- dfm(toks, remove = stopwords())
mt <- dfm_trim(mt, min_termfreq = 10)
lss <- textmodel_lss(mt, seedwords('pos-neg'))
summary(lss)

# sentiment model on economy
eco <- head(char_keyness(toks, 'econom*'), 500)
lss_eco <- textmodel_lss(mt, seedwords('pos-neg'), features = eco)

# sentiment model on politics
pol <- head(char_keyness(toks, 'politi*'), 500)
lss_pol <- textmodel_lss(mt, seedwords('pos-neg'), features = pol)
}
\references{
Watanabe, Kohei. "Measuring News Bias: Russia's Official News
  Agency ITAR-TASS' Coverage of the Ukraine Crisis." European Journal of
  Communication 32, no. 3 (March 20, 2017): 224–41.
  https://doi.org/10.1177/0267323117695735.
}
