% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{textmodel_lss}
\alias{textmodel_lss}
\title{A vector-space model for subject specific sentiment-analysis}
\usage{
textmodel_lss(
  x,
  seeds,
  features = NULL,
  k = 300,
  weight = "count",
  cache = FALSE,
  simil_method = "cosine",
  include_data = TRUE,
  engine = c("RSpectra", "rsvd", "irlba", "text2vec"),
  s = k,
  w = 50,
  d = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a dfm or fcm created by \code{\link[quanteda:dfm]{quanteda::dfm()}} or
\code{\link[quanteda:fcm]{quanteda::fcm()}}}

\item{seeds}{a character vector, named numeric vector or dictionary that
contains seed words.}

\item{features}{features of a dfm to be included in the model as terms. This
argument is used to make models only sensitive to subject specific words.}

\item{k}{the size of semantic space passed to the SVD engine Only used when
\code{x} is a \code{dfm}.}

\item{weight}{weighting scheme passed to \code{\link[quanteda:dfm_weight]{quanteda::dfm_weight()}}.
Ignored when \code{engine} is "text2vec".}

\item{cache}{if \code{TRUE}, save retult of SVD for next execution with
identical \code{x} and \code{k}.}

\item{simil_method}{specifies method to compute similarity between features.
The value is passed to \code{\link[quanteda:textstat_simil]{quanteda::textstat_simil()}}, "cosine" is
used otherwise.}

\item{include_data}{if \code{TRUE}, fitted model include the dfm supplied as
\code{x}.}

\item{engine}{choose SVD engine between \code{\link[RSpectra:svds]{RSpectra::svds()}},
\code{\link[irlba:irlba]{irlba::irlba()}}, and \code{\link[text2vec:GlobalVectors]{text2vec::GlobalVectors()}}.}

\item{s}{the number factors used to compute similarity between features.}

\item{w}{the size of word vectors. Only used when \code{engine} is
"text2vec".}

\item{d}{eigen value value weighting. Only used when \code{engine} is
"RSpectra" or "irlba".}

\item{verbose}{show messages if \code{TRUE}.}

\item{...}{additional argument passed to the SVD engine}
}
\description{
A vector-space model for subject specific sentiment-analysis
}
\examples{
\dontrun{
require(quanteda)

# Available at https://bit.ly/2GZwLcN
corp <- readRDS("data_corpus_guardian2016-10k.rds")
corp <- corpus_reshape(data_corpus_guardian, 'sentences')
toks <- tokens(corp, remove_punct = TRUE) \%>\%
        tokens_remove(stopwords()) \%>\%
dfmat <- dfm(toks) \%>\%
         dfm_trim(dfmat, min_termfreq = 10)

# SVD
lss <- textmodel_lss(dfmat, seedwords('pos-neg'))
summary(lss)

# sentiment model on economy
eco <- head(char_keyness(toks, 'econom*'), 500)
lss_eco <- textmodel_lss(dfmat, seedwords('pos-neg'), features = eco)

# sentiment model on politics
pol <- head(char_keyness(toks, 'politi*'), 500)
lss_pol <- textmodel_lss(dfmat, seedwords('pos-neg'), features = pol)

# GloVe
toks <- tokens_select("^[\\\\p{L}]+$", valuetype = "regex", padding = TRUE)
fcmat  <- fcm(toks, context = "window", count = "weighted", weights = 1 / (1:5), tri = TRUE)
lss <- textmodel_lss(fcmat, seedwords('pos-neg'))
}
}
\references{
Watanabe, Kohei. "Measuring News Bias: Russia's Official News
Agency ITAR-TASS' Coverage of the Ukraine Crisis." European Journal of
Communication 32, no. 3 (March 20, 2017): 224–41.
https://doi.org/10.1177/0267323117695735.
}
