% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot_simil}
\alias{textplot_simil}
\title{Plot similarity between seed words}
\usage{
textplot_simil(x)
}
\arguments{
\item{x}{fitted textmodel_lss object.}
}
\description{
Plot similarity between seed words
}
