% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{textplot_simil}
\alias{textplot_simil}
\title{Plot similarity of seed words}
\usage{
textplot_simil(x, group = FALSE)
}
\arguments{
\item{x}{fitted textmodel_lss object}

\item{group}{if \code{TRUE} group seed words by seed patterns and show
average similarity}
}
\description{
Plot similarity of seed words
}
