% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot_terms}
\alias{textplot_terms}
\title{Plot polarity scores of words}
\usage{
textplot_terms(x, highlighted = NULL, max_highlighted = 50, max_words = 10000)
}
\arguments{
\item{x}{a fitted textmodel_lss object.}

\item{highlighted}{\link[quanteda:pattern]{quanteda::pattern} to select words to highlight.}

\item{max_highlighted}{the maximum number of words to highlight. When
\code{highlighted = NULL}, words to highlight are randomly selected proportionally
to \code{polarity ^ 2 * log(frequency)}.}

\item{max_words}{the maximum number of words to plot. Words are randomly
sampled to keep the number below the limit.}
}
\description{
Plot polarity scores of words
}
