% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lss.R
\name{weight_seeds}
\alias{weight_seeds}
\title{Internal function to generate equally-weighted seed set}
\usage{
weight_seeds(seeds, type)
}
\description{
Internal function to generate equally-weighted seed set
}
\keyword{internal}
